requirements.txt
praw==7.7.1
pymongo==4.6.1
python-dotenv==1.0.0
fastapi
uvicorn[standard]
pydantic
cryptography
sentence-transformers>=3.0.0 